moved loading the site grades fragment into the course sites fragment file
